# docs/api/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyFomin API Documentation",
        "description": "API documentation for MyFomin",
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get all transactions with categories and accounts",
                "description": "Returns a list of all transactions ordered by creation date (descending)",
                "operationId": "5235e2842a72577d9671d351ad6d9456",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved transaction list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Create a new transaction (deposit or spend)",
                "description": "Adds a new transaction and updates the account balance.",
                "operationId": "011c7bd4e5cf22aa760e58118f74bdcc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "account_id",
                                    "category_id",
                                    "amount"
                                ],
                                "properties": {
                                    "type": {
                                        "description": "Transaction type",
                                        "type": "string",
                                        "enum": [
                                            "deposit",
                                            "withdrawal"
                                        ],
                                        "example": "deposit"
                                    },
                                    "account_id": {
                                        "description": "Account ID",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category_id": {
                                        "description": "Category ID",
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "amount": {
                                        "description": "Transaction amount",
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500.5
                                    },
                                    "description": {
                                        "description": "Transaction description (optional)",
                                        "type": "string",
                                        "example": "Deposit from client"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Transaction"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error creating transaction",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transactions/{id}": {
            "put": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Update a transaction",
                "description": "Update a transaction.",
                "operationId": "1d97935fb3ce3aced1e1fa1226e5c21f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Transaction ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransactionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransactionDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request data"
                    },
                    "404": {
                        "description": "Transaction not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Delete a transaction",
                "description": "Delete a transaction.",
                "operationId": "5067d94301b6d899832b47b99862a4a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Transaction ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Transaction deleted successfully"
                    },
                    "404": {
                        "description": "Transaction not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AccountRequest": {
                "title": "AccountRequest",
                "description": "Account model schema",
                "required": [
                    "name",
                    "amount"
                ],
                "properties": {
                    "name": {
                        "description": "* @OA\\Schema(\n    schema=\"AccountRequest\",\n    type=\"object\",\n    title=\"AccountRequest\",\n    description=\"Account model schema\",\n    required={\"name\", \"amount\"},",
                        "type": "string",
                        "example": "Main Account"
                    },
                    "amount": {
                        "type": "number",
                        "example": 5000
                    }
                },
                "type": "object"
            },
...
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Electronics"
                    },
                    "type": {
                        "type": "string",
                        "example": "product"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "title": "ErrorResponse",
                "description": "error response interface",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "short description"
                    },
                    "detail": {
                        "type": "string",
                        "example": "detail"
                    }
                },
                "type": "object"
            },
            "TransactionDto": {
                "required": [
                    "id",
                    "account_id",
                    "type",
                    "amount",
                    "description",
                    "category_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "account_id": {
                        "type": "integer"
                    },
                    "type": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "integer"
                    },
                    "description": {
                        "type": "string"
                    },
                    "category_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Transaction": {
                "title": "Transaction",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "deposit",
                            "withdrawal"
                        ],
                        "example": "deposit"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 500.75
                    },
                    "description": {
                        "type": "string",
                        "example": "Payment received"
                    },
                    "account_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-06T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-06T12:00:00Z"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "account": {
                        "$ref": "#/components/schemas/AccountRequest"
                    }
                },
                "type": "object"
            }
        }
    },
...
}